phishing-detector/requirements.txt
Flask==2.0.1
python-whois==0.8.0
requests==2.26.0
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.4